feat: add support for time-to-live (TTL) (#65)
This PR aims to add a support for time-to-live (TTL) on Squid. This means that
entries (sentences) can expire after a certain time (expressed in seconds). This
can reduce storage, but increase RAM usage (due to the indexes required).
This PR also refound the stockage system. From now, everything is not stocked
in ONE file, but splitted into multiple files. This should improve scalability.

Signed-off-by: RealHinome <hinome@gravitalia.com>